requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-texttospeech==2.16.3
pydantic==2.5.0
python-multipart==0.0.6